refactor: MenuGroup JPA 리팩토링
