+++ src/main/java/kitchenpos/application/menugroup/MenuGroupService.java
@@ -1,8 +1,8 @@
 package kitchenpos.application.menugroup;
 
 import kitchenpos.application.menugroup.dto.MenuGroupCreateRequest;
-import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
+import kitchenpos.repository.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,20 +11,21 @@ import java.util.List;
 @Service
 public class MenuGroupService {
 
-    private final MenuGroupDao menuGroupDao;
+    private final MenuGroupRepository menuGroupRepository;
 
-    public MenuGroupService(final MenuGroupDao menuGroupDao) {
-        this.menuGroupDao = menuGroupDao;
+    public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+        this.menuGroupRepository = menuGroupRepository;
     }
 
     @Transactional
-    public MenuGroup create(final MenuGroupCreateRequest menuGroupCreateRequest) {
-        MenuGroup menuGroup = new MenuGroup(menuGroupCreateRequest.getName());
+    public MenuGroup create(final MenuGroupCreateRequest req) {
+        MenuGroup menuGroup = new MenuGroup(null, req.getName());
 
-        return menuGroupDao.save(menuGroup);
+        return menuGroupRepository.save(menuGroup);
     }
 
+    @Transactional(readOnly = true)
     public List<MenuGroup> list() {
-        return menuGroupDao.findAll();
+        return menuGroupRepository.findAll();
     }
 }
