+++ src/main/java/kitchenpos/domain/MenuGroup.java
@@ -1,10 +1,22 @@
 package kitchenpos.domain;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.util.Objects;
 
+@Entity
+@Table(name = "MENU_GROUP")
 public class MenuGroup {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(nullable = false)
     private String name;
 
     protected MenuGroup() {
@@ -12,10 +24,6 @@ public class MenuGroup {
 
     public MenuGroup(final Long id, final String name) {
         this.id = id;
-        this.name = name;
-    }
-
-    public MenuGroup(final String name) {
         this.name = name;
     }
