+++ src/main/java/kitchenpos/repository/MenuGroupRepository.java
@@ -1,11 +1,13 @@
-package kitchenpos.dao;
+package kitchenpos.repository;
 
 import kitchenpos.domain.MenuGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface MenuGroupDao {
+public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
+
     MenuGroup save(MenuGroup entity);
 
     Optional<MenuGroup> findById(Long id);
