+++ src/test/java/kitchenpos/fixture/MenuGroupFixture.java
@@ -6,10 +6,10 @@ import kitchenpos.domain.MenuGroup;
 public class MenuGroupFixture {
 
     public static MenuGroup 메뉴_그룹_생성() {
-        return new MenuGroup("메뉴그룹");
+        return new MenuGroup(null, "메뉴그룹");
     }
 
     public static MenuGroup 메뉴_그룹_생성(final String name) {
-        return new MenuGroup(name);
+        return new MenuGroup(null, name);
     }
 }
